requirements.txt
flask
qrcode
pillow
cryptography
requests
pytest
opencv-python
pyzbar
pyserial
